fix int conversion

+++ BodoSQL/bodosql/tests/test_kernels/test_casting_array_kernels.py
@@ -406,7 +406,7 @@ def test_cast_int8(numeric_arrays, memory_leak_check):
             # we normally must calculate the underflow/overflow behavior for ints
             # however casting to float64 first will handle this for us
             if f_elem.is_integer():
-                return np.int8(elem)
+                return np.int64(elem).astype(np.int8)
             else:
                 return np.int8(_round_float(f_elem))
